src/main/java/com/leetcode/concurrency/BoundedBlockingQueue.java
package com.leetcode.concurrency;

import java.util.LinkedList;
import java.util.concurrent.locks.Condition;
import java.util.concurrent.locks.ReentrantLock;

/**
 * @author 阿尔曼
 */
public class BoundedBlockingQueue {

    private final int capacity;
    private final LinkedList<Integer> queue;
    private final ReentrantLock lock;
    private final Condition empty;
    private final Condition full;

    public BoundedBlockingQueue(int capacity) {
        this.capacity = capacity;
        this.queue = new LinkedList<>();
        this.lock = new ReentrantLock();
        this.empty = this.lock.newCondition();
        this.full = this.lock.newCondition();
    }

    public void enqueue(int element) throws InterruptedException {
        lock.lock();
        try {
            while (queue.size() == capacity) {
                full.await();
            }
            queue.add(element);
            empty.signal();
        } finally {
            lock.unlock();
        }
    }

    public int dequeue() throws InterruptedException {
        lock.lock();
        try {
            while (queue.size() == 0) {
                empty.await();
            }
            int first = queue.pop();
            full.signal();
            return first;
        } finally {
            lock.unlock();
        }
    }

    public int size() {
        lock.lock();
        try {
            return queue.size();
        } finally {
            lock.unlock();
        }
    }
}
